css style nav, header
Gave each one a display flex and left the container alone (body) since it's by default responsive
Started the card component and I need to make an overlay, so was thinking
of using grid 1/-1 to place it ontop of each other.
However, with the up and down animation prob absolute would be a better way with overflow?