Fix .empty-message position
Since it shared the same div with grid, so later down the line my grid are going to be pushed down since empty-message is still in the html flow.
Also, we could make it absolute to not distrub the grid, but then we would need to give grid-container a height, which we don't want to do since grid will populate it through auto.